test(executor): split context injection tests into dedicated file
