+++ runtime/src/llm/chat-executor.test.ts
@@ -3205,188 +3205,6 @@ describe("ChatExecutor", () => {
   });
 
   // --------------------------------------------------------------------------
-  // Injection
-  // --------------------------------------------------------------------------
-
-  describe("injection", () => {
-    it("skillInjector.inject() result appears in messages", async () => {
-      const skillInjector = {
-        inject: vi.fn().mockResolvedValue("Skill context: you can search"),
-      };
-      const provider = createMockProvider();
-      const executor = new ChatExecutor({
-        providers: [provider],
-        skillInjector,
-      });
-
-      await executor.execute(createParams());
-
-      const messages = (provider.chat as ReturnType<typeof vi.fn>).mock
-        .calls[0][0] as LLMMessage[];
-      expect(messages[1]).toMatchObject({
-        role: "system",
-        content: "Skill context: you can search",
-      });
-    });
-
-    it("skillInjector failure is non-blocking", async () => {
-      const skillInjector = {
-        inject: vi.fn().mockRejectedValue(new Error("injection failed")),
-      };
-      const provider = createMockProvider();
-      const executor = new ChatExecutor({
-        providers: [provider],
-        skillInjector,
-      });
-
-      const result = await executor.execute(createParams());
-
-      expect(result.content).toBe("mock response");
-      const messages = (provider.chat as ReturnType<typeof vi.fn>).mock
-        .calls[0][0] as LLMMessage[];
-      // Only system prompt + user message (no skill context)
-      expect(messages).toHaveLength(2);
-    });
-
-    it("memoryRetriever failure is non-blocking", async () => {
-      const memoryRetriever = {
-        retrieve: vi.fn().mockRejectedValue(new Error("retrieval failed")),
-      };
-      const provider = createMockProvider();
-      const executor = new ChatExecutor({
-        providers: [provider],
-        memoryRetriever,
-      });
-
-      const result = await executor.execute(createParams());
-
-      expect(result.content).toBe("mock response");
-      const messages = (provider.chat as ReturnType<typeof vi.fn>).mock
-        .calls[0][0] as LLMMessage[];
-      expect(messages).toHaveLength(2);
-    });
-
-    it("memoryRetriever.retrieve() result appears in messages", async () => {
-      const memoryRetriever = {
-        retrieve: vi
-          .fn()
-          .mockResolvedValue("Memory: user prefers short answers"),
-      };
-      const provider = createMockProvider();
-      const executor = new ChatExecutor({
-        providers: [provider],
-        memoryRetriever,
-      });
-
-      await executor.execute(
-        createParams({
-          history: [{ role: "assistant", content: "Previous turn" }],
-        }),
-      );
-
-      const messages = (provider.chat as ReturnType<typeof vi.fn>).mock
-        .calls[0][0] as LLMMessage[];
-      expect(messages[1]).toMatchObject({
-        role: "system",
-        content: "Memory: user prefers short answers",
-      });
-    });
-
-    it("progressProvider.retrieve() result appears in messages", async () => {
-      const progressProvider = {
-        retrieve: vi
-          .fn()
-          .mockResolvedValue("## Recent Progress\n\n- [tool_result] ran ls"),
-      };
-      const provider = createMockProvider();
-      const executor = new ChatExecutor({
-        providers: [provider],
-        progressProvider,
-      });
-
-      await executor.execute(
-        createParams({
-          history: [{ role: "assistant", content: "Previous turn" }],
-        }),
-      );
-
-      const messages = (provider.chat as ReturnType<typeof vi.fn>).mock
-        .calls[0][0] as LLMMessage[];
-      expect(messages).toEqual(
-        expect.arrayContaining([
-          expect.objectContaining({
-            role: "system",
-            content: "## Recent Progress\n\n- [tool_result] ran ls",
-          }),
-        ]),
-      );
-    });
-
-    it("progressProvider failure is non-blocking", async () => {
-      const progressProvider = {
-        retrieve: vi.fn().mockRejectedValue(new Error("progress backend down")),
-      };
-      const provider = createMockProvider();
-      const executor = new ChatExecutor({
-        providers: [provider],
-        progressProvider,
-      });
-
-      const result = await executor.execute(createParams());
-
-      expect(result.content).toBe("mock response");
-    });
-
-    it("progressProvider injected after learningProvider", async () => {
-      const learningProvider = {
-        retrieve: vi.fn().mockResolvedValue("## Learned Patterns\n\n- lesson"),
-      };
-      const progressProvider = {
-        retrieve: vi.fn().mockResolvedValue("## Recent Progress\n\n- step"),
-      };
-      const provider = createMockProvider();
-      const executor = new ChatExecutor({
-        providers: [provider],
-        learningProvider,
-        progressProvider,
-      });
-
-      await executor.execute(
-        createParams({
-          history: [{ role: "assistant", content: "Previous turn" }],
-        }),
-      );
-
-      const messages = (provider.chat as ReturnType<typeof vi.fn>).mock
-        .calls[0][0] as LLMMessage[];
-      // System prompt + learning + progress = 3 system messages before history/user
-      const systemMessages = messages.filter((m) => m.role === "system");
-      expect(systemMessages).toHaveLength(3);
-      expect(systemMessages[1].content).toContain("Learned Patterns");
-      expect(systemMessages[2].content).toContain("Recent Progress");
-    });
-
-    it("does not inject persistent memory providers on a fresh session", async () => {
-      const memoryRetriever = { retrieve: vi.fn().mockResolvedValue("Memory") };
-      const learningProvider = { retrieve: vi.fn().mockResolvedValue("Learning") };
-      const progressProvider = { retrieve: vi.fn().mockResolvedValue("Progress") };
-      const provider = createMockProvider();
-      const executor = new ChatExecutor({
-        providers: [provider],
-        memoryRetriever,
-        learningProvider,
-        progressProvider,
-      });
-
-      await executor.execute(createParams({ history: [] }));
-
-      expect(memoryRetriever.retrieve).not.toHaveBeenCalled();
-      expect(learningProvider.retrieve).not.toHaveBeenCalled();
-      expect(progressProvider.retrieve).not.toHaveBeenCalled();
-    });
-  });
-
-  // --------------------------------------------------------------------------
   // Per-call streaming
   // --------------------------------------------------------------------------
